requirement.txt
# Core data science libraries
pandas>=1.5.0
numpy>=1.21.0
scikit-learn>=1.2.0
scipy>=1.9.0

# Machine learning models
xgboost>=1.7.0
lightgbm>=3.3.0
catboost>=1.1.0

# Visualization
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.10.0
shap>=0.41.0

# Model interpretability
lime>=0.2.0.1
eli5>=0.13.0

# Hyperparameter optimization
optuna>=3.0.0
hyperopt>=0.2.7

# Data processing
imbalanced-learn>=0.9.0
category-encoders>=2.5.0
feature-engine>=1.5.0

# Utilities
tqdm>=4.64.0
joblib>=1.2.0
pyyaml>=6.0
click>=8.0.0

# Jupyter notebook
jupyter>=1.0.0
ipykernel>=6.15.0
ipywidgets>=7.7.0

# Testing
pytest>=7.0.0
pytest-cov>=4.0.0

# Documentation
sphinx>=5.0.0
sphinx-rtd-theme>=1.0.0

# Development
black>=22.0.0
flake8>=5.0.0
pre-commit>=2.20.0

# Environment management
python-dotenv>=0.20.0

# For downloading UCI dataset
ucimlrepo>=0.0.3